+++ posthog-ai/src/index.ts
@@ -2,8 +2,10 @@ import PostHogOpenAI from './openai'
 import PostHogAzureOpenAI from './openai/azure'
 import { wrapVercelLanguageModel } from './vercel/middleware'
 import PostHogAnthropic from './anthropic'
+import { LangChainCallbackHandler } from './langchain/callbacks'
 
 export { PostHogOpenAI as OpenAI }
 export { PostHogAzureOpenAI as AzureOpenAI }
 export { PostHogAnthropic as Anthropic }
 export { wrapVercelLanguageModel as withTracing }
+export { LangChainCallbackHandler }