+++ posthog-ai/CHANGELOG.md
@@ -1,4 +1,18 @@
-# Next
+# 3.2.1 - 2025-02-11
+
+- fix: add experimental_wrapLanguageModel to vercel middleware supporting older versions of ai
+
+# 3.2.0 - 2025-02-11
+
+- feat: change how we handle streaming support for openai and anthropic
+
+# 3.1.1 - 2025-02-07
+
+- fix: bump ai to 4.1.0
+
+# 3.1.0 - 2025-02-07
+
+- feat: add posthogCostOverride, posthogModelOverride, and posthogProviderOverride to sendEventToPosthog for vercel
 
 # 2.4.0 - 2025-02-03